Add files via upload
The app allows user to add applications in a list and runs all of them with a single click. Useful in cases where user regularly runs a set of applications which now he can do with a single click of a button as this python app remembers those set of applications for the user.